requirements.txt
black
click
clickhouse-connect
psycopg2-binary
pymongo[srv]
